EduContextFlow/requirements.txt
google-genai
pillow
flask
flask-cors
